Added the code and data
The purpose of the exercise was to analyse data about education in Chicago. The analysis was done in two parts: one is a Jupyter notebook (SQL_Chicago_Data.ipynb), and the other is an SQL file (Coursera_test_answers.sql), with the associated function Coursera_test_function.sql. Both parts work with the same database, which is included as CSV files.